# backend/felcn-base-backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend Base NestJS
  version: '1.0'
  description: Documentación Backend base - NestJS
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'https://base.agcs.agetic.gob.bo/ws/api'
paths:
...
  /autorizacion/roles:
    get:
      operationId: RolController_listar
      summary: API para obtener el listado de Roles
      parameters: []
      responses:
        '200':
          description: ''
      tags: &ref_2
        - Roles
      security: &ref_3
        - bearer: []
    post:
      operationId: RolController_crear
      summary: API para crear un rol
      parameters: []
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearRolDto'
      responses:
        '201':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/todos:
    get:
      operationId: RolController_listarTodos
      summary: API para obtener el listado de todos los Roles
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/{id}:
    patch:
      operationId: RolController_actualizar
      summary: API para actualizar un rol
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarRolDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/{id}/activacion:
    patch:
      operationId: RolController_activar
      summary: API para activar un rol
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/{id}/inactivacion:
    patch:
      operationId: RolController_inactivar
      summary: API para inactivar un rol
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
...
  /token:
    post:
      operationId: RefreshTokensController_getAccessToken
      summary: API que actualiza el token de acceso usando un refresh token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenDto'
      responses:
        '201':
          description: ''
      tags: &ref_7
        - Refresh Token
  /{id}:
    delete:
      operationId: RefreshTokensController_eliminarRefreshToken
      summary: API que elimina el refresh token
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_7
      security:
        - bearer: []
...
  /parametros:
    get:
      operationId: ParametroController_listar
      summary: API para obtener el listado de parámetros
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_9
        - Parámetros
      security: &ref_10
        - bearer: []
    post:
      operationId: ParametroController_crear
      summary: API para crear un nuevo parámetro
      parameters: []
      requestBody:
        required: true
        description: new Parametro
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearParametroDto'
      responses:
        '201':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{grupo}/listado:
    get:
      operationId: ParametroController_listarPorGrupo
      summary: API para obtener el listado de parámetros por grupo
      parameters:
        - name: grupo
          required: true
          in: path
          example: TD
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{id}:
    patch:
      operationId: ParametroController_actualizar
      summary: API para actualizar un parámetro
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarParametroDto'
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{id}/activacion:
    patch:
      operationId: ParametroController_activar
      summary: API para activar un parámetro
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{id}/inactivacion:
    patch:
      operationId: ParametroController_inactivar
      summary: API para inactivar un parámetro
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CrearRolDto:
      type: object
      properties:
        rol:
          type: string
          example: CONSULTA
        nombre:
          type: string
          example: Consulta
        estado:
          type: string
          example: ACTIVO
      required:
        - rol
        - nombre
        - estado
    ActualizarRolDto:
      type: object
      properties:
        rol:
          type: string
          example: ADMINISTRADOR
        nombre:
          type: string
          example: Administrador
        estado:
          type: string
          example: ACTIVO
      required:
        - rol
        - nombre
        - estado
...
    TokenDto:
      type: object
      properties:
        token:
          type: string
          example: ''
      required:
        - token
...
    CrearParametroDto:
      type: object
      properties:
        codigo:
          type: string
          example: TD-CI
        nombre:
          type: string
          example: Cédula de identidad
        grupo:
          type: string
          example: CD
        descripcion:
          type: string
          example: Cédula de identidad
        estado:
          type: string
          example: ACTIVO
      required:
        - codigo
        - nombre
        - grupo
        - descripcion
        - estado
    ActualizarParametroDto:
      type: object
      properties:
        codigo:
          type: string
          example: TD-CI
        nombre:
          type: string
          example: Cédula de identidad
        grupo:
          type: string
          example: CD
        descripcion:
          type: string
          example: Cédula de identidad
        estado:
          type: string
          example: ACTIVO
      required:
        - codigo
        - nombre
        - grupo
        - descripcion
        - estado